\name{bestsol}
\alias{bestsol}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Best solution monitoring function}
\description{
Display best solution from result of GA
}
\usage{
bestsol(garesult)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{garesult}{GA result object}
}
\value{
Display chromosome, fitness value and generation number of best solution.
}
\author{
Zeynel Cebeci & Erkut Tekeli
}
