\name{fixpcmut}
\alias{fixpcmut}
\title{
Static crossover and mutation rate
}
\description{
The function is used when the crossover and mutation rates are not changed throughout the GA run.
}
\usage{
fixpcmut(cxpc, mutpm, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cxpc}{Crossover rate}
  \item{mutpm}{Mutation rate}
  \item{\dots}{Further arguments passed to or from other methods.}
}
\value{
  \item{pc}{Crossover rate}
  \item{pm}{Mutation rate}
}
\author{
Zeynel Cebeci & Erkut Tekeli
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 \code{\link{ilmdhc}},
 \code{\link{adana1}},
 \code{\link{adana2}},
 \code{\link{leitingzhi}},
 \code{\link{adana3}}
}
